refactor: extract boundary expansion logic to shared chunk_range module
