+++ tests/unit/research/test_chunk_range.py
@@ -1,16 +1,13 @@
 """Tests for smart boundary expansion edge cases."""
 
-from chunkhound.services.deep_research_service import DeepResearchService
-
-
-def _service() -> DeepResearchService:
-    return DeepResearchService.__new__(DeepResearchService)
+from chunkhound.services.research.shared.chunk_range import (
+    expand_to_natural_boundaries,
+)
 
 
 def test_boundary_expansion_skips_invalid_start() -> None:
-    service = _service()
     lines = ["line"] * 5
-    start, end = service._expand_to_natural_boundaries(
+    start, end = expand_to_natural_boundaries(
         lines=lines,
         start_line=10,
         end_line=12,
@@ -21,9 +18,8 @@ def test_boundary_expansion_skips_invalid_start() -> None:
 
 
 def test_boundary_expansion_skips_invalid_end() -> None:
-    service = _service()
     lines = ["line"] * 5
-    start, end = service._expand_to_natural_boundaries(
+    start, end = expand_to_natural_boundaries(
         lines=lines,
         start_line=1,
         end_line=0,
@@ -34,9 +30,8 @@ def test_boundary_expansion_skips_invalid_end() -> None:
 
 
 def test_boundary_expansion_skips_reversed_range() -> None:
-    service = _service()
     lines = ["line"] * 5
-    start, end = service._expand_to_natural_boundaries(
+    start, end = expand_to_natural_boundaries(
         lines=lines,
         start_line=4,
         end_line=2,
